server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
    # via flask
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
    # via flask
flask==3.0.0
    # via -r requirements.txt
flask-peewee==3.0.4
huey==2.5.0
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
kombu==5.3.4
lxml == 4.9.3
    # via -r requirements.txt
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
numpy==1.26.2
    # via -r requirements.txt
prompt-toolkit==3.0.41
pandas==2.1.4
    # via -r requirements.txt
peewee==3.17.0
python-barcode==0.15.1
    # via -r requirements.txt
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
svg-stack==0.1.0
    # via -r requirements.txt
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
werkzeug==3.0.1
    # via flask
wtf-peewee==3.0.5
wtforms==3.1.1
